LayoutTests/http/tests/mime/html-with-html-expected.txt

CONSOLE MESSAGE: Pass

html-document.html has MIME type text/html
